requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.21.0
numpy